implemented stats apis, modified over the app, and aslo home page's sections

+++ src/lib/DateFormatter.ts
@@ -1,4 +1,4 @@
 export const formatDate = (dateString: string) => {
-    const options: Intl.DateTimeFormatOptions = { year: "numeric", month: "long", day: "numeric" };
+    const options: Intl.DateTimeFormatOptions = { year: "numeric", month: "short", day: "numeric" };
     return new Date(dateString).toLocaleDateString(undefined, options);
 }
